OnlineShopProject
User
This application allows user to view different products according to the
the categories,adding to cart and checking out with the details of the
customer.This application also shows the available number of items of
all the products.All the validations are made wherever found necessary.
After placing an order with the customer details,I have designed the
application to add the customer details and order details in the
database and generate a Transaction number automatically,which will be
shown to user.
Admin
This module consists of Admin Login.This allows the admin to add new
category or products to the database.Provided the feature of uploading
an Image of the product
Tracking  Details
This allows users to view their order according to the Transaction
Number.Displays the customer and order information.